Makes sure the previously selected gene is expanded when another matching pathway is selected while the 'temporary' left-drawer is closed -- #227
